+++ targets/pic_8bit/mikroe/pic18/src/uart/implementation_2/hal_ll_uart.c
@@ -793,7 +793,7 @@ void hal_ll_uart_irq_disable( handle_t *handle, hal_ll_uart_irq_t irq ) {
     }
 }
 
-void hal_ll_uart_write( handle_t *handle, uint16_t wr_data) {
+void hal_ll_uart_write( handle_t *handle, uint8_t wr_data) {
     const hal_ll_uart_base_handle_t *hal_ll_hw_reg = hal_ll_uart_get_base_handle;
 
     while( !check_reg_bit( hal_ll_hw_reg->uart_uerrir_reg_addr, HAL_LL_UART_TXMTIF_BIT ) );
@@ -801,7 +801,7 @@ void hal_ll_uart_write( handle_t *handle, uint16_t wr_data) {
     write_reg( hal_ll_hw_reg->uart_txb_reg_addr, wr_data );
 }
 
-uint16_t hal_ll_uart_read( handle_t *handle ) {
+uint8_t hal_ll_uart_read( handle_t *handle ) {
     const hal_ll_uart_base_handle_t *hal_ll_hw_reg = hal_ll_uart_get_base_handle;
 
     return read_reg( hal_ll_hw_reg->uart_rxb_reg_addr );
